MA-GCL-main/requirements.txt
# alabaster==0.7.12
# alembic==1.5.7
# anaconda-client==1.7.2
# anaconda-navigator==1.10.0
# anaconda-project==0.8.3
appdirs==1.4.4
argh==0.26.2
argon2-cffi==20.1.0
asn1crypto==1.4.0
astor==0.8.1
astroid==2.4.2
astropy==4.0.2
async-generator==1.10
atomicwrites==1.4.0
attrs==20.3.0
autopep8==1.6.0
Babel==2.8.1
backcall==0.2.0
backports.functools-lru-cache==1.6.4
backports.shutil-get-terminal-size==1.0.0
backports.tempfile==1.0
backports.weakref==1.0.post1
beautifulsoup4==4.9.3
bitarray==1.6.1
bkcharts==0.2
bleach==3.2.1
bokeh==2.2.3
boto==2.49.0
Bottleneck==1.3.2
brotlipy==0.7.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cliff==3.7.0
cloudpickle==1.6.0
clyent==1.2.2
cmaes==0.8.2
cmd2==1.5.0
colorama==0.4.4
colorlog==4.7.2
conda==4.12.0
conda-build==3.20.5
conda-package-handling==1.7.3
conda-verify==3.4.2
contextlib2==0.6.0.post1
cryptography==3.1.1
cycler==0.10.0
Cython==0.29.21
cytoolz==0.11.0
dask==2.30.0
decorator==4.4.2
defusedxml==0.6.0
dgl==0.8.0.post1
dgl-cu100==0.5.2
dgl-cu101==0.8a211204
dgl-cu110==0.6.1
dgl-cu111==0.6.0.post1
diff-match-patch==20200713
distlib==0.3.1
distributed==2.30.1
docutils==0.16
entrypoints==0.3
et-xmlfile==1.0.1
fastcache==1.1.0
filelock==3.0.12
flake8==3.8.4
Flask==1.1.2
fsspec==0.8.3
future==0.18.2
gensim==3.8.3
gevent==20.9.0
glob2==0.7
gmpy2==2.0.8
googledrivedownloader==0.4
greenlet==0.4.17
h5py==2.10.0
HeapDict==1.0.1
html5lib==1.1
hyperopt==0.1.2
idna==2.10
imageio==2.9.0
imagesize==1.2.0
importlib-metadata==2.0.0
iniconfig==1.1.1
intervaltree==3.1.0
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isodate==0.6.0
isort==5.10.1
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.17.1
jeepney==0.5.0
Jinja2==2.11.2
joblib==0.17.0
json-tricks==3.15.5
json5==0.9.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
jupyterlab==2.2.6
jupyterlab-pygments==0.1.2
jupyterlab-server==1.2.0
keras==2.8.0
keyring==21.4.0
kiwisolver==1.3.0
lazy-object-proxy==1.4.3
libarchive-c==2.9
littleutils==0.2.2
llvmlite==0.34.0
locket==0.2.0
lxml==4.6.1
Mako==1.1.4
MarkupSafe==1.1.1
matplotlib==3.3.2
mccabe==0.6.1
memory-profiler==0.58.0
mistune==0.8.4
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
mock==4.0.2
more-itertools==8.6.0
mpmath==1.1.0
msgpack==1.0.0
multipledispatch==0.6.0
munkres==1.1.4
navigator-updater==0.2.1
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.2
networkx==2.5
nltk==3.5
nose==1.3.7
notebook==6.1.4
numba==0.51.2
numexpr==2.7.1
numpy==1.19.2
numpydoc==1.1.0
ogb==1.3.1
olefile==0.46
openpyxl==3.0.5
optuna==2.6.0
outdated==0.2.1
packaging==20.4
pandas==1.1.3
pandocfilters==1.4.3
parso==0.7.0
partd==1.1.0
path==15.0.0
pathlib2==2.3.5
pathtools==0.1.2
patsy==0.5.1
pbr==5.5.1
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
pip==20.2.4
pkginfo==1.6.1
pluggy==0.13.1
ply==3.11
prettytable==2.1.0
prometheus-client==0.8.0
prompt-toolkit==3.0.8
protobuf==3.19.4
psutil==5.7.2
ptyprocess==0.6.0
py==1.9.0
pycodestyle==2.8.0
pycosat==0.6.3
pycparser==2.20
pycurl==7.43.0.6
pydocstyle==5.1.1
pyflakes==2.2.0
PyGCL==0.1.1
Pygments==2.7.2
pylint==2.6.0
pymongo==4.0.1
pyodbc==4.0.0
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyperclip==1.8.2
pyrsistent==0.17.3
PySocks==1.7.1
pytest==0.0.0
python-dateutil==2.8.1
python-editor==1.0.4
python-jsonrpc-server==0.4.0
python-language-server==0.35.1
python-louvain==0.15
PythonWebHDFS==0.2.3
pytz==2020.1
PyWavelets==1.1.1
pyxdg==0.27
PyYAML==5.3.1
pyzmq==19.0.2
QDarkStyle==2.8.1
QtAwesome==1.0.1
qtconsole==4.7.7
QtPy==1.9.0
rdflib==6.0.0
regex==2020.10.15
requests==2.24.0
responses==0.16.0
rope==0.18.0
Rtree==0.9.4
ruamel-yaml==0.15.87
ruamel.yaml.clib==0.2.6
schema==0.7.5
scikit-image==0.17.2
scikit-learn==1.0.1
scipy==1.5.2
seaborn==0.11.0
SecretStorage==3.1.2
Send2Trash==1.5.0
setuptools==50.3.1.post20201107
simplegeneric==0.8.1
simplejson==3.17.6
singledispatch==3.4.0.3
sip==4.19.13
six==1.15.0
smart-open==4.1.0
snowballstemmer==2.0.0
sortedcollections==1.2.1
sortedcontainers==2.2.2
soupsieve==2.0.1
Sphinx==3.2.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sphinxcontrib-websupport==1.2.4
spyder==4.1.5
spyder-kernels==1.9.4
SQLAlchemy==1.3.20
statsmodels==0.12.0
stevedore==3.3.0
sympy==1.6.2
tables==3.6.1
tabulate==0.8.7
tblib==1.7.0
tensorboardX==2.5
terminado==0.9.1
testpath==0.4.4
threadpoolctl==2.1.0
tifffile==2020.10.1
toml==0.10.1
toolz==0.11.1
torch==1.9.0+cu111
torch-cluster==1.5.9
torch-geometric==2.0.3
torch-scatter==2.0.9
torch-sparse==0.6.12
torch-spline-conv==1.2.1
torchaudio==0.9.0
torchvision==0.10.0+cu111
tornado==6.0.4
tqdm==4.50.2
traitlets==5.0.5
tsnecuda==3.0.1
typing-extensions==3.7.4.3
ujson==4.0.1
unicodecsv==0.14.1
urllib3==1.25.11
virtualenv==20.4.0
watchdog==0.10.3
wcwidth==0.2.5
webencodings==0.5.1
websockets==10.1
Werkzeug==1.0.1
wheel==0.35.1
widgetsnbextension==3.5.1
wrapt==1.11.2
wurlitzer==2.0.1
xlrd==1.2.0
XlsxWriter==1.3.7
xlwt==1.3.0
xmltodict==0.12.0
yacs==0.1.8
yapf==0.30.0
zict==2.0.0
zipp==3.4.0
zope.event==4.5.0
zope.interface==5.1.2